# openapi-oficina.yaml
openapi: 3.0.3
info:
  title: API Oficina Mecânica
  version: "v1"
  description: |
    API em Python (Flask) para gestão de autenticação de usuários (JWT), busca externa
    de produtos (token de serviço) e carrinho por usuário. Documentação em PT-BR.
  contact:
    name: Suporte da API
    email: suporte@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Desenvolvimento local
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Autenticação]
      summary: Registrar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, email, senha]
              properties:
                nome: { type: string }
                email: { type: string, format: email }
                senha: { type: string, minLength: 6 }
            examples:
              ex:
                value: { nome: "João", email: "joao@ex.com", senha: "123456" }
      responses:
        '201': { description: Criado }
        '400': { $ref: '#/components/responses/Resp400' }
        '409': { $ref: '#/components/responses/Resp409' }
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Login e emissão de JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha]
              properties:
                email: { type: string, format: email }
                senha: { type: string }
            examples:
              ex:
                value: { email: "joao@ex.com", senha: "123456" }
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/UsuarioPublico' }
        '400': { $ref: '#/components/responses/Resp400' }
        '401': { $ref: '#/components/responses/Resp401' }
  /auth/me:
    get:
      tags: [Autenticação]
      summary: Dados do usuário autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioPublico' }
        '401': { $ref: '#/components/responses/Resp401' }
    put:
      tags: [Autenticação]
      summary: Atualizar perfil do usuário
      description: Atualiza nome, telefone e/ou senha (usar `senha_atual` + `nova_senha`).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: { type: string }
                telefone: { type: string }
                senha_atual: { type: string }
                nova_senha: { type: string, minLength: 6 }
      responses:
        '200':
          description: Atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioPublico' }
        '400': { $ref: '#/components/responses/Resp400' }
        '401': { $ref: '#/components/responses/Resp401' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Resp400: { description: Requisição inválida }
    Resp401: { description: Não autorizado }
...
    Resp409: { description: Conflito (duplicidade) }
...
  schemas:
    UsuarioPublico:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string, format: email }
        telefone: { type: string }